Allow native browser zoom
Users with impaired vision might need to increase the size of the map and its
controls. Not allowing them to zoom in is an innecessary accessibility hurdle.
I assume it has been blocked to prevent a mix-up with the map zoom. However,
since the Zoom buttons get bigger when you use the browser zoom, I would argue
that users should be able to notice them and recover from their mistake quickly.

Committed on behalf of DB Systel